Tidy intermediate_representation tests Part I
